multi_agent_search/requirements.txt
deepagents
langgraph
langchain
langchain-community
langchain-text-splitters
langchain-mcp-adapters
langchain-google-genai
langchain-ollama
langchain-groq
langchain-chroma>=0.1.0
rank-bm25  # BM25 for keyword-based search
chromadb>=0.4.0
mcp
python-dotenv
pydantic
pydantic-ai  # Type-safe AI agent framework
pydantic-ai-slim[openai,groq]  # Providers for Pydantic AI
httpx
watchdog
supabase
tavily-python
fastapi
uvicorn[standard]
PyJWT[crypto]
stripe
inngest

# RAG Dependencies
sentence-transformers>=2.2.0
pypdf
python-docx>=1.1.0
docx2txt
unstructured[pdf]

# Local multimodal (optional)
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0

